refactor: split protected example controller
